Add define for use with hid-remapper's drag scroll
Added #define to choose what far right button is on layer 0. This is
useful for drag scroll on https://github.com/jfedor2/hid-remapper.
(hid-remapper supports resolution multiplier, this is one way to hold
myself over...)

I should have exported the JSON config for the remapper, but I'll do
that another time. I might take Ploopy home to work on it. Or I might
leave it at work so that I don't spend all my free time working on it.

Is there another way to ensure I don't spend all my free time (and the
time I should be spending asleep) coding? This gets to be a problem, and
I don't know the solution.

Oh this is a commit message, not a diary. Sorry.